+++ src/test/java/wooteco/subway/controller/StationControllerUnitTest.java
@@ -23,7 +23,6 @@ import org.springframework.test.web.servlet.MockMvc;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.service.StationService;
-import wooteco.subway.ui.StationController;
 
 @WebMvcTest(StationController.class)
 class StationControllerUnitTest {
